ingestion_pipeline_weather/requirements.txt
requests==2.32.3
pandas==2.2.2
google-cloud-bigquery==3.25.0
pyarrow==16.1.0
pytz